CI Cleanup

See merge request radar/radarkit!84

+++ source/RKHealthRelayNaveen.c
@@ -54,7 +54,7 @@ static int healthRelayNaveenRead(RKClient *client) {
 static int healthRelayNaveenGreet(RKClient *client) {
     ssize_t s = RKNetworkSendPackets(client->sd, "\n", 1, NULL);
     if (s < 0) {
-        return -s;
+        return (int)-s;
     }
     RKLog("%s Naveen @ %s:%d connected.\n", client->name, client->hostIP, client->port);
     return RKResultSuccess;
